EditarPublicacion OK, falta que cargue el tipo de publicacion

+++ WindowsFormsApplication1/DataProvider/DBMapper.cs
@@ -219,7 +219,7 @@ namespace MercadoEnvio
 
         /** Publicaciones **/
 
-        public Publicacion ObtenerPublicacion(Decimal idPublicacion)
+        public Publicacion ObtenerPublicacion(int idPublicacion)
         {
             Publicacion objeto = new Publicacion();
             Type clase = objeto.GetType();
